--- src/NServiceBus.Extensions.DependencyInjection/ContainerSettings.cs
@@ -1,40 +1,0 @@
-﻿namespace NServiceBus
-{
-    using System;
-    using System.Collections.Generic;
-    using Microsoft.Extensions.DependencyInjection;
-
-    /// <summary>
-    /// Settings for the provided dependency injection container.
-    /// </summary>
-    [ObsoleteEx(Message = "Use the externally managed container mode to integrate with third party dependency injection containers.",
-        RemoveInVersion = "3",
-        TreatAsErrorFromVersion = "2")]
-    public class ContainerSettings<TContainerBuilder>
-    {
-        /// <summary>
-        /// The <see cref="IServiceCollection"/> used to be used by the <see cref="IServiceProvider"/>.
-        /// </summary>
-        [ObsoleteEx(Message = "Use the externally managed container mode to integrate with third party dependency injection containers.",
-            RemoveInVersion = "3",
-            TreatAsErrorFromVersion = "2")]
-        public IServiceCollection ServiceCollection => throw new NotImplementedException();
-
-        /// <summary>
-        /// Provides access to container specific configuration as part of creating the <see cref="IServiceProvider"/>.
-        /// </summary>
-        [ObsoleteEx(Message = "Use the externally managed container mode to integrate with third party dependency injection containers.",
-            RemoveInVersion = "3",
-            TreatAsErrorFromVersion = "2")]
-        public void ConfigureContainer(Action<TContainerBuilder> containerConfiguration)
-        {
-            throw new NotImplementedException();
-        }
-
-        internal List<Action<TContainerBuilder>> ContainerConfigurations { get; } = new List<Action<TContainerBuilder>>(0);
-
-        internal ContainerSettings()
-        {
-        }
-    }
-}
