--- src/NServiceBus.Extensions.DependencyInjection/Guard.cs
@@ -1,15 +1,0 @@
-﻿namespace NServiceBus
-{
-    using System;
-
-    static class Guard
-    {
-        public static void AgainstNull(string argumentName, object value)
-        {
-            if (value == null)
-            {
-                throw new ArgumentNullException(argumentName);
-            }
-        }
-    }
-}
