--- src/NServiceBus.Extensions.DependencyInjection/IStartableEndpointWithExternallyManagedContainerExtensions.cs
@@ -1,25 +1,0 @@
-﻿namespace NServiceBus
-{
-    using System;
-    using System.Threading.Tasks;
-
-    /// <summary>
-    /// An extension point to create an endpoint in the start-up phase where the container is externally managed using the Microsoft dependency injection abstraction.
-    /// </summary>
-    public static class IStartableEndpointWithExternallyManagedContainerExtensions
-    {
-        /// <summary>
-        /// Starts the endpoint and returns a reference to it.
-        /// </summary>
-        /// <param name="startableEndpoint">The startable endpoint.</param>
-        /// <param name="serviceProvider">The container specific service provider factory.</param>
-        /// <returns>A reference to the endpoint.</returns>
-        public static Task<IEndpointInstance> Start(this IStartableEndpointWithExternallyManagedContainer startableEndpoint, IServiceProvider serviceProvider)
-        {
-            Guard.AgainstNull(nameof(startableEndpoint), startableEndpoint);
-            Guard.AgainstNull(nameof(serviceProvider), serviceProvider);
-
-            return startableEndpoint.Start(serviceProvider);
-        }
-    }
-}
