+++ src/NServiceBus.Extensions.DependencyInjection/obsoletes.cs
@@ -4,6 +4,11 @@
     using System.Collections.Generic;
     using Microsoft.Extensions.DependencyInjection;
 
+    // An internal type referenced by the API approvals test as it can't reference obsoleted types.
+    class InternalType
+    {
+    }
+
     /// <summary>
     /// Extension methods to integrate containers support the Microsoft.Extensions.DependencyInjection model.
     /// </summary>
@@ -57,6 +62,41 @@
 
         internal ContainerSettings()
         {
+        }
+    }
+
+    /// <summary>
+    /// An extension point to create an endpoint in the start-up phase where the container is externally managed using the Microsoft dependency injection abstraction.
+    /// </summary>
+    [ObsoleteEx(
+        ReplacementTypeOrMember = nameof(IStartableEndpointWithExternallyManagedContainer),
+        RemoveInVersion = "3",
+        TreatAsErrorFromVersion = "2")]
+    public static class IStartableEndpointWithExternallyManagedContainerExtensions
+    {
+    }
+
+    /// <summary>
+    /// Provides factory methods for creating endpoint instances with an externally managed container implementing the Microsoft dependency injection abstraction.
+    /// </summary>
+    [ObsoleteEx(
+        ReplacementTypeOrMember = nameof(EndpointWithExternallyManagedContainer),
+        RemoveInVersion = "4",
+        TreatAsErrorFromVersion = "3")]
+    public static class EndpointWithExternallyManagedServiceProvider
+    {
+        /// <summary>
+        /// Creates a new startable endpoint based on the provided configuration that uses an externally managed container implementing the Microsoft dependency injection abstraction.
+        /// </summary>
+        /// <param name="endpointConfiguration">The endpoint configuration.</param>
+        /// <param name="serviceCollection">Service collection.</param>
+        [ObsoleteEx(
+            Message = "NServiceBus supports Microsoft.Extensions.DependencyInjection directly. Change usage of `EndpointWithExternallyManagedServiceProvider` to `EndpointWithExternallyManagedContainer` and remove the reference to the NServiceBus.Extensions.DependencyInjection package",
+            RemoveInVersion = "3",
+            TreatAsErrorFromVersion = "2")]
+        public static IStartableEndpointWithExternallyManagedContainer Create(EndpointConfiguration endpointConfiguration, IServiceCollection serviceCollection)
+        {
+            throw new NotImplementedException();
         }
     }
 }
